+++ packages/apis/example/lib/services/handlers/discount_handlers/discount_code_handler/retrieves_count_of_discount_codes_handler.dart
@@ -1,5 +1,5 @@
 import 'package:apis/apis.dart';
-import 'package:apis/network/remote/discounts/discount_code/abstract/discount_code.dart';
+import 'package:apis/network/remote/discounts/discount_code/abstract/discount_code_service.dart';
 import 'package:apis/network/remote/discounts/discount_code/freezed_model/response/discount_code_count_response.dart';
 import 'package:example/services/api_request_handler.dart';
 import 'package:get_it/get_it.dart';
